Add communication based on Unix domain sockets
Introduce first version of MazeRunner API with basic that
external Runners can use to interact with simulator.

API can be used to test algorithms for searching through the maze
and calculating speedrun routes.

For now requests are just printed out. Error is
returned in response, indicating that requests are not
implemented.

Signed-off-by: Witold Lipieta <witek103@gmail.com>